Refactor: Split monolithic CanvasArea into modular components and hooks (#89)
* Refactor: split monolithic CanvasArea into modular components and hooks

* refactor(Bubble): split into modular sub-components

- Extract useDraggableWithBounds hook to src/hooks/
- Create icons/ folder with ScribbleIcon and TrapezoidIcon
- Create utils/ folder with richTextUtils (applyStyleToRichText, etc.)
- Split UI into 5 section components:
  - BubbleCollapsed: collapsed state indicator
  - BubbleTextSection: fonts, sizes, formatting, alignment
  - BubbleShapeSection: geometric shape selector
  - BubbleStrokeSection: dash style, size, color, opacity
  - BubbleFillSection: fill color and opacity

Reduces Bubble.tsx from ~1448 to ~740 lines (~50% reduction)
Part of #58
