+++ src/components/Bubble/icons/index.ts
@@ -1,0 +1,2 @@
+export { ScribbleIcon } from './ScribbleIcon';
+export { TrapezoidIcon } from './TrapezoidIcon';
