+++ src/components/Bubble/sections/index.ts
@@ -1,0 +1,5 @@
+export { BubbleCollapsed } from './BubbleCollapsed';
+export { BubbleTextSection } from './BubbleTextSection';
+export { BubbleShapeSection } from './BubbleShapeSection';
+export { BubbleStrokeSection } from './BubbleStrokeSection';
+export { BubbleFillSection } from './BubbleFillSection';
